Refactor to move request stuff out of config
